+++ aws/spec/doc_test.go
@@ -17,7 +17,7 @@ func TestDocForEachCommand(t *testing.T) {
 }
 func TestDocForEachParam(t *testing.T) {
 	for name, def := range AWSTemplatesDefinitions {
-		params, opts := params.List(def.Params)
+		params, opts, _ := params.List(def.Params)
 		for _, param := range append(params, opts...) {
 			if doc, ok := awsdoc.TemplateParamsDoc(def.Action, def.Entity, param); !ok || doc == "" {
 				t.Fatalf("missing documentation for param '%s' for '%s'", param, name)